@@ -1,6 +1,6 @@
 [metadata]
 name = tikzplotlib
-version = 0.9.12
+version = 0.9.14
 author = Nico Schlömer
 author_email = nico.schloemer@gmail.com
 description = Convert matplotlib figures into TikZ/PGFPlots
@@ -14,16 +14,16 @@ long_description_content_type = text/markdown
 license = MIT
 license_file = LICENSE
 classifiers =
-    Development Status :: 4 - Beta
+    Development Status :: 5 - Production/Stable
     Framework :: Matplotlib
     License :: OSI Approved :: MIT License
     Operating System :: OS Independent
     Programming Language :: Python
     Programming Language :: Python :: 3
-    Programming Language :: Python :: 3.6
     Programming Language :: Python :: 3.7
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
+    Programming Language :: Python :: 3.10
     Topic :: Multimedia :: Graphics :: Graphics Conversion
     Topic :: Scientific/Engineering :: Visualization
 keywords =
@@ -41,7 +41,7 @@ install_requires =
     matplotlib >= 1.4.0
     numpy
     Pillow
-python_requires = >=3.6
+python_requires = >=3.7
 
 [options.packages.find]
 where=src
